Added Track Number + 1.scpt and Track Number +/- x.scpt
These are two scripts I wrote and use with iTunes.  Track Number +
1.scpt adds 1 to the track number of each selected song.  This is
useful, for example, if you joined songs and want to renumber the tracks
in your album so that the joined tracks are integrated with the track
numbers.  It runs with no interaction, though it beeps when done.  Do
not use iTunes while it is running.  Also, be sure that your songs are
sorted by something that will not change when the track number changes,
such as Track Number or Disk Number.  Sorting by Album is fine.

Track Number +/- x.scpt is similar, except it asks you for an amount to
add or subtract each track number by.

Put these in /Library/iTunes/Scripts/ or ~/Library/iTunes/Scripts/ to
put them in the iTunes script menu.

Sorry for the unicode filename.  Seriously, though, if you don't have
Mac OS X, this isn't going to do anything for you anyway.  Besides, any
modern operating system should be able to handle unicode in a filename.