Added Dashboard Starter.app
This is maybe not the best way to do this, but it works.  It loads the
Dashboard for you.  Put Dashboard Starter.app in your login items if you
don't want to wait for the Dashboard to load the first time you access
it.

Requires Mac OS X.  And if you really want this feature, try one of the
freeware applications out there that do this much better.