Added git-backup.sh
This is a simple shell script that commits changes in a git repository,
then immediately reverts them, serving to "Backup" the unstaged changes
for 30 days.

This lets you edit the commit message so you can use it to help you find
the SHA1 to cherry-pick later on. It uses commit -a, so it probably
would have unintended consequences if you have changes staged but not
committed (i.e., it would essentially unstage them).