Add Comics Downloader.py
This is a script that I started writing a while ago when ACMEReader
broke for comics.com, but stopped writing when it was fixed.  It doesn't
work again, so I am working on it again.  I am pretty sure that it
doesn't work yet.

See the docstring of the file for more information.